refactor: extract multi-selection from entity-selection state
Decomposes the entity-selection.svelte.ts state into a standalone multi-selection module handling advanced across logic. Part of issue #628.
